app/requirement.txt
python-dotenv
fastapi
langchain
langchain_core
langchain_community
langchain_openai
langchain-qdrant
sentence_transformers
pypdf
